moves functions to db
